REFACTOR :: AttendanceCalculator, CertificateCalculator, VolunteerCalculatro 훅에 비즈니스 로직, 뷰에 합쳤습니다
